Addition of the PrimeDynamic object.
The PrimeDynamic object is essentially a dynamic prime tester; When
it tests whether a number is prime or not, it will first check if it
has ever tested that prime before. If so, it will use the result it returned
last time. If it has not yet run into that number, then it will manually test
it and then save the result, and then return it.

This was a long overdue addition to the eratosthenes prime module.

-Zach